feat(mongoengine): add DataRetentionDocument base class
